backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
email-validator
pydantic
openai